% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_geno_individual}
\alias{check_geno_individual}
\title{Check genotypic representation.}
\usage{
check_geno_individual(x)
}
\arguments{
\item{x}{List. Individual}
}
\description{
Check if the genotypic representation is valid.
}
