% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_pos}
\alias{check_pos}
\title{Check if genetic positions are valid}
\usage{
check_pos(x)
}
\arguments{
\item{x}{Genetic positions to be checked.}
}
\description{
Check if genetic positions are valid
}
