% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_xodat_individual}
\alias{check_xodat_individual}
\title{Check if crossover-data are valid}
\usage{
check_xodat_individual(x)
}
\arguments{
\item{x}{Crossover data to be checked}
}
\description{
Check if crossover-data are valid
}
