% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Converter$convert}
\alias{Converter$convert}
\title{convert}
\arguments{
\item{individual}{List. An individual to be converted.}
}
\description{
Convert an individual from its segmental to its genotypic representation.
}
