% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Converter$insert_founder}
\alias{Converter$insert_founder}
\title{insert_founder}
\arguments{
\item{keys}{Integer vector. The two (unique) founder alleles.}

\item{geno}{List. The genotypic data.}
}
\description{
Insert founder genotypes.
}
