% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xo_founder.R
\name{create_xo_founder}
\alias{create_xo_founder}
\title{Create founders with crossover-data.}
\usage{
create_xo_founder(alleles, L)
}
\arguments{
\item{alleles}{Integer vector of length 2, giving founder alleles.}

\item{L}{Double vector. Chromosome lengths in cM.}
}
\value{
A nested list. The first level refers to gametes, the second
to chromosomes.
}
\description{
Create founders with crossover-data.
}
\examples{
create_xo_founder(alleles = c(-5L, 54L), L = c(32.2, 65.3, 88.2))

}
\seealso{
\code{\link{cross_xo}}, \code{\link{meiosis_xo}}
}
