% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{cross_geno}
\alias{cross_geno}
\alias{cross_geno}
\alias{dh_geno}
\title{Cross individuals (genotypic representation)}
\usage{
cross_geno(father, mother, positions, xoparam)

dh_geno(individual, positions, xoparam)
}
\arguments{
\item{father}{List. Father.}

\item{mother}{List. Father.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}
}
\value{
List. A new individual.
}
