% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{cross_xo}
\alias{cross_xo}
\alias{cross_xo}
\alias{dh_xo}
\title{Cross individuals (segmental representation)}
\usage{
cross_xo(father, mother, xoparam)

dh_xo(individual, xoparam)
}
\arguments{
\item{father}{List. Father.}

\item{mother}{List. Father.}

\item{xoparam}{List. Crossover parameters.}

\item{positions}{List. Genetic positions.}
}
\value{
List. A new individual.
}
