% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R
\docType{package}
\name{Meiosis}
\alias{Meiosis}
\alias{Meiosis-package}
\title{Meiosis: A package for the simulation of meiosis events in diploids.}
\description{
This packages provides tools for the simulation of meiosis events in diploid species.
}
