% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{meiosis_geno}
\alias{meiosis_geno}
\alias{meiosis_geno}
\title{Simulate meiosis (genotypic representation)}
\usage{
meiosis_geno(individual, positions, xoparam)
}
\arguments{
\item{individual}{List. Individual.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}
}
\value{
List. A new gamete.
}
