% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{meiosis_xo}
\alias{meiosis_xo}
\alias{meiosis_xo}
\title{Simulate meiosis (segmental representation)}
\usage{
meiosis_xo(individual, xoparam, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new gamete.
}
\description{
Simulation of meiosis events.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::meiosis_xo(exdat$founder, exdat$xoparam)

}
