% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{seed_rng}
\alias{seed_rng}
\alias{seed_rng}
\title{Seed the random number generator.}
\usage{
seed_rng(seed = NULL)
}
\arguments{
\item{seed}{Integer (default = \code{NULL}). If \code{NULL}, a random seed is used.}
}
\value{
Integer. The used seed.
}
\description{
A Mersenne Twister pseudo-random generator of 32-bit numbers with a state size of
10037 bits is used.
}
\examples{
Meiosis::seed_rng(123L)
the_seed <- Meiosis::seed_rng()

}
