% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{self_geno}
\alias{self_geno}
\alias{self_geno}
\title{Produce selfing (genotypic representation)}
\usage{
self_geno(individual, positions, xoparam, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new individual.
}
\description{
Simulation of a selfing.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::self_geno(exdat$ind, exdat$positions, exdat$xoparam)

}
