% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R
\name{self_xo}
\alias{self_xo}
\title{Produce selfing (segmental representation)}
\usage{
self_xo(individual, xoparam, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new individual.
}
\description{
Simulation of a selfing.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::self_xo(exdat$founder, exdat$xoparam)

}
