% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_positions}
\alias{check_positions}
\title{Check genetic positions}
\usage{
check_positions(x)
}
\arguments{
\item{x}{List. Genetic positions.}
}
\description{
Check if genetic positions are valid.
}
