% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_xoparam}
\alias{check_xoparam}
\title{Check crossover parameters.}
\usage{
check_xoparam(x)
}
\arguments{
\item{x}{List. Crossover parameters.}
}
\value{
\code{Null}
}
\description{
Check if crossover parameters are valid.
}
