% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{cross_geno}
\alias{cross_geno}
\alias{cross_geno}
\title{Cross individuals (genotypic representation)}
\usage{
cross_geno(father, mother, positions, xoparam, check = FALSE)
}
\arguments{
\item{father}{List. Father.}

\item{mother}{List. Mother.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new individual.
}
\description{
Simulation of a cross.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::cross_geno(exdat$ind, exdat$ind2, exdat$positions, exdat$xoparam)

}
