% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{cross_xo}
\alias{cross_xo}
\alias{cross_xo}
\title{Cross individuals (segmental representation)}
\usage{
cross_xo(father, mother, xoparam, check = FALSE)
}
\arguments{
\item{father}{List. Father.}

\item{mother}{List. Mother.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new individual.
}
\description{
Simulation of a cross.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::cross_xo(exdat$founder, exdat$founder2, exdat$xoparam)

}
