% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{crossover}
\alias{crossover}
\alias{crossover}
\title{Simulate crossover locations.}
\usage{
crossover(L, m, p, obligate_chiasma, Lstar)
}
\arguments{
\item{L}{Double. Length of the chromosome in cM. Must be > 50.}

\item{m}{Integer. Interference parameter for chi-square model. Must be non-negative.}

\item{p}{Double. Proportion of chiasmata coming from no-interference process}

\item{obligate_chiasma}{Logical. Should a least on the 4-strand bundle at meiosis.}

\item{Lstar}{Double. Reduced chromosome length as produced by \code{\link{calc_Lstar}}.}
}
\value{
Numeric Vector. Crossover locations.
}
\details{
This function is an R-wrapper of an underlying C++ routine.
It is not intended for direct usage, but exposed for completeness.
}
\examples{
crossover(300, 10, 0.5, FALSE, 300)

}
\seealso{
\code{\link[simcross]{sim_crossover}}
}
