% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R
\name{dh_geno}
\alias{dh_geno}
\title{Produce doubled haploid (genotypic representation)}
\usage{
dh_geno(individual, positions, xoparam, use_names = FALSE, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}

\item{use_names}{Logical. Should loci names be preserved if present?}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new individual.
}
\description{
Simulation of a doubled haploid.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::dh_geno(exdat$ind, exdat$positions, exdat$xoparam)

}
