% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{dh_geno}
\alias{dh_geno}
\alias{dh_geno}
\title{Produce doubled haploid (genotypic representation)}
\usage{
dh_geno(individual, positions, xoparam)
}
\arguments{
\item{individual}{List. Individual.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}
}
\value{
List. A new individual.
}
\description{
Simulation of a doubled haploid.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::dh_geno(exdat$ind, exdat$positions, exdat$xoparam)

}
