% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R
\name{dh_xo}
\alias{dh_xo}
\title{Produce doubled haploid (segmental representation)}
\arguments{
\item{individual}{List. Individual.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}
}
\value{
List. A new individual.
}
