% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R
\name{dh_xo}
\alias{dh_xo}
\title{Produce doubled haploid (segmental representation)}
\usage{
dh_xo(individual, xoparam, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new individual.
}
\description{
Simulation of a doubled haploid.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::dh_xo(exdat$founder, exdat$xoparam)

}
