% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R
\docType{package}
\name{Meiosis}
\alias{Meiosis}
\alias{Meiosis-package}
\title{Meiosis: Simulation of meiosis in plant breeding research}
\description{
This packages provides tools for the simulation of meiosis in plant breeding research.
}
