% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{meiosis_geno}
\alias{meiosis_geno}
\alias{meiosis_geno}
\title{Simulate meiosis (genotypic representation)}
\usage{
meiosis_geno(individual, positions, xoparam, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{positions}{List. Genetic positions.}

\item{xoparam}{List. Crossover parameters.}

\item{check}{Logical. Should checks be performed?}
}
\value{
List. A new gamete.
}
\description{
Simulation of meiosis events.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::meiosis_geno(exdat$ind, exdat$positions, exdat$xoparam)

}
