% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{meiosis_xo}
\alias{meiosis_xo}
\alias{meiosis_xo}
\title{Simulate meiosis (segmental representation)}
\usage{
meiosis_xo(individual, xoparam)
}
\arguments{
\item{individual}{List. Individual.}

\item{xoparam}{List. Crossover parameters.}
}
\value{
List. A new gamete.
}
\description{
Simulation of meiosis events.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::meiosis_xo(exdat$founder, exdat$xoparam)

}
