% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{realized_coancestry}
\alias{realized_coancestry}
\alias{realized_coancestry}
\title{Compute the realized coefficient of co-ancestry between two individuals.}
\usage{
realized_coancestry(individual_1, individual_2 = NULL)
}
\arguments{
\item{individual_1}{List. A first Individual.}

\item{individual_2}{List. A second Individual. If \code{NULL}, self-relationship of
\code{individual_1} is computed.}
}
\value{
double. Realized coefficient of co-ancestry.
}
\details{
The realized coefficient of co-ancestry is (herein) defined as the probability
that two alleles randomly drawn from one of the homologous chromosomes, each from one of the
individuals, are identical-by-descent (as specified by origins of founder alleles).
}
