% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/exports.R
\name{realized_heter}
\alias{realized_heter}
\alias{realized_heter}
\title{Compute realized heterozygosity.}
\usage{
realized_heter(individual, check = FALSE)
}
\arguments{
\item{individual}{List. Individual.}

\item{check}{Logical. Should checks be performed?}
}
\value{
double. Realized heterozygosity.
}
\description{
Compute the realized heterozygosity of an individual, i.e., the proportion of
the genome where founder alleles match.
}
\examples{
data(exdat, package = 'Meiosis')
Meiosis::realized_heter(Meiosis::cross_xo(exdat$founder, exdat$founder, exdat$xoparam))

}
