% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meiosis.R, R/RcppExports.R
\name{seed_rng}
\alias{seed_rng}
\alias{seed_rng}
\title{Seed the random number generator of the C++ routines.}
\usage{
seed_rng(seed = NULL)
}
\arguments{
\item{seed}{Integer. If \code{NULL}, a random seed is used.}
}
\value{
Integer. The seed used.
}
\description{
A Mersenne Twister pseudo-random generator of 32-bit numbers
with a state size of 10037 bits is used in the underlying C++ routines.
}
\details{
The state of the random number generator used by the C++ routines is not
shared with the \code{R} session from which these routines are interfaced.
}
\examples{
Meiosis::seed_rng(123L)
the_seed <- Meiosis::seed_rng()

}
